优化ToDoList

+++ app/src/main/java/com/example/myapplication/TaskAdapter.java
@@ -1,6 +1,8 @@
 package com.example.myapplication;
 
 import android.annotation.SuppressLint;
+import android.app.AlertDialog;
+import android.content.DialogInterface;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -45,13 +47,29 @@ public class TaskAdapter extends RecyclerView.Adapter<TaskAdapter.TaskViewHolder
         holder.cbCompleted.setOnCheckedChangeListener((buttonView, isChecked) -> {
             task.setCompleted(isChecked);
             dbHelper.updateTaskCompletion(task.getId(), isChecked);
+
         });
 
         holder.itemView.setOnLongClickListener(v -> {
-            int taskId = task.getId();
-            tasks.remove(position);
-            notifyItemRemoved(position);
-            dbHelper.deleteTask(taskId);
+//            int taskId = task.getId();
+//            tasks.remove(position);
+//            notifyItemRemoved(position);
+//            dbHelper.deleteTask(taskId);
+            // 用户点击对话框外部，关闭对话框
+            new AlertDialog.Builder(v.getContext())
+                    .setTitle("确认删除")
+                    .setMessage("您确定要删除" + task.getTitle() + "吗?")
+                    .setPositiveButton("删除", (dialog, which) -> {
+                        int taskId = task.getId();
+                        tasks.remove(position);
+                        notifyItemRemoved(position); // 更新 RecyclerView
+                        dbHelper.deleteTask(taskId);
+                    })
+                    .setNegativeButton("取消", (dialog, which) -> {
+                        dialog.dismiss();
+                    })
+                    .setOnCancelListener(DialogInterface::dismiss)
+                    .show();
             return true;
         });
     }
